changed proposal model to go through the new model, still have to alter it so that invitations can also go through it
